requirements.txt
ollama
pandas
numpy==1.22.4
langchain
matplotlib
dotenv
pinecone-client
pinecone
django
tiktoken
opencv-python==4.5.5.64
pytesseract
flask
flask-cors
pypdf
pdfplumber
python-docx
nltk